wym-project/pyproject.toml
[tool.poetry]
name = "wym-project"
version = "0.1.0"
description = "c'est notre projet"
authors = ["Alex <alexandra.cecile.penin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
